Update: Refactored code, log out errors

+++ schemas/test.py
@@ -1,11 +1,10 @@
 from pydantic import BaseModel, Field
-from decimal import Decimal
 
 
 class Test(BaseModel):
     name: str = Field(min_length=2, max_length=50, description="The test name")
     description: str | None
-    price: Decimal = Field(description="The test price")
+    price: int = Field(description="The test price")
 
 
 class TestInDB(Test):
